+++ keyoscacquire/docs/changelog.md
@@ -1,4 +1,6 @@
 # Changelog
 
-- v1.0.1: Added license, changes in README, tiny changes in some syntax
+NEED TO INCORPORATE NEW FUNC IN README
+
+- v1.0.1: New command line function for obtaining a set number of traces consecutively, Added license, changes in README
 - v1.0.0: First release